Fix routing

+++ src/todolist/todolist.service.ts
@@ -28,11 +28,11 @@ export class TodosService {
         throw new HttpException('Todo not found', HttpStatus.NOT_FOUND);
     }
 
-    async createTodo(todo: CreateTodoDto) {
+    async createTodo(userId: string, todo: CreateTodoDto) {
         const newTodo = await this.todoRepository.create(todo);
         await this.todoRepository.save(newTodo);
 
-        return this.todoRepository.find();
+        return this.getUserTodos(userId);
     }
 
     async updateTodo(userId: string, id: number, todo: UpdateTodoDto) {
@@ -44,17 +44,22 @@ export class TodosService {
             const updatedTodo = await this.todoRepository.findOne({
                 where: { id: id },
             });
-            if (updatedTodo) return this.todoRepository.find();
+            if (updatedTodo) return this.getUserTodos(userId);
         }
 
         throw new HttpException('Todo not found', HttpStatus.NOT_FOUND);
     }
 
-    async deleteTodo(id: number) {
-        const deletedTodo = await this.todoRepository.delete(id);
-        if (!deletedTodo.affected) {
-            throw new HttpException('Todo not found', HttpStatus.NOT_FOUND);
+    async deleteTodo(userId: string, id: number) {
+        const targetTodo = await this.todoRepository.findOne({
+            where: { user: userId, id },
+        });
+        if (targetTodo) {
+            const deletedTodo = await this.todoRepository.delete(id);
+            if (deletedTodo.affected) {
+                return this.getUserTodos(userId);
+            }
         }
-        return this.todoRepository.find();
+        throw new HttpException('Todo not found', HttpStatus.NOT_FOUND);
     }
 }
